# packages/mtmai/mtmai-0.3.1259-py3-none-any.whl/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.1259"
  },
  "paths": {
...
    "/api/v1/tasks/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Task List",
        "operationId": "tasks-task_list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/{mttask_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Mttask",
        "description": "Get task by id.",
        "operationId": "tasks-get_mttask",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mttask_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mttask Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MtTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/task_types": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Task Types",
        "description": "Get task types.",
        "operationId": "tasks-task_types",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectItemsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/mttask_update_status": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Mttask Update Status",
        "description": "\u66f4\u65b0\u4efb\u52a1\u72b6\u6001",
        "operationId": "tasks-mttask_update_status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MttaskUpdateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/schedule/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Schedule Get",
        "operationId": "tasks-schedule_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDetailPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/schedule/list": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Schedule List",
        "description": "Retrieve site auto items with pagination.",
        "operationId": "tasks-schedule_list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_SiteAutoItemPublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/schedule/create": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Schedule Create",
        "operationId": "tasks-schedule_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/schedule/update/{id}": {
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Schedule Update",
        "operationId": "tasks-schedule_update",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/chat_profile/{id}": {
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Cht Profile Delete",
        "description": "\u6839\u636e\u8c03\u5ea6\u914d\u7f6e\uff0c\u7acb\u5373\u751f\u6210\u4e00\u4e2a\u8fd0\u884c\u4efb\u52a1",
        "operationId": "tasks-cht_profile_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommonResultResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          },
          "data": {
            "type": "object",
            "title": "Data",
            "default": {}
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "CommonResultResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListResponse_SiteAutoItemPublic_": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SiteAutoItemPublic"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "ListResponse[SiteAutoItemPublic]"
      },
...
      "MtTask": {
        "properties": {
          "schedule_id": {
            "type": "string",
            "format": "uuid",
            "title": "Schedule Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "$ref": "#/components/schemas/MtTaskStatus",
            "default": "new"
          },
          "is_auto": {
            "type": "boolean",
            "title": "Is Auto",
            "default": false
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 3
          },
          "inputs": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputs",
            "description": "\u8f93\u5165\u53c2\u6570",
            "default": {}
          },
          "state": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "\u4efb\u52a1\u72b6\u6001",
            "default": {}
          },
          "outputs": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs",
            "description": "\u8f93\u51fa\u7ed3\u679c",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "execution_state": {
            "type": "string",
            "title": "Execution State",
            "description": "(\u672a\u542f\u7528)\u6267\u884c\u72b6\u6001,manual_review_required \u8868\u793a\u9700\u8981\u4eba\u5de5\u5e72\u9884",
            "default": ""
          },
          "task_timeout": {
            "type": "integer",
            "title": "Task Timeout",
            "description": "(\u672a\u542f\u7528)\u4efb\u52a1\u8d85\u65f6\u65f6\u95f4, \u5355\u4f4d\u79d2, \u5982\u679c\u4efb\u52a1\u5728\u89c4\u5b9a\u65f6\u95f4\u5185\u6ca1\u6709\u5b8c\u6210\uff0c\u5219\u8ba4\u4e3a\u4efb\u52a1\u5931\u8d25",
            "default": 86400
          },
          "heap_timeout": {
            "type": "integer",
            "title": "Heap Timeout",
            "description": "(\u672a\u542f\u7528)\u5fc3\u8df3\u8d85\u65f6\u65f6\u95f4, \u5355\u4f4d\u79d2, \u5982\u679c\u4efb\u52a1\u5728\u89c4\u5b9a\u65f6\u95f4\u5185\u6ca1\u6709\u53d1\u9001\u5fc3\u8df3\uff0c\u5219\u8ba4\u4e3a\u4efb\u52a1\u5931\u8d25",
            "default": 300
          },
          "heaptbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heaptbeat At",
            "description": "(\u672a\u542f\u7528)\u5fc3\u8df3\u65f6\u95f4, \u6700\u8fd1\u4e00\u8c03\u5ea6\uff0c\u6216\u8005\u6b65\u9aa4\u975e\u5e38\u591a\u8fd0\u884c\u65f6\u95f4\u8f83\u957f\u7684\u4efb\u52a1\uff0c\u53ef\u4ee5\u5728\u5de5\u4f5c\u8fd0\u884c\u8fc7\u7a0b\u4e2d\uff0c\u53d1\u9001\u5fc3\u8df3\u5305\uff0c\u8868\u793a\u4efb\u52a1\u8fd8\u5728\u6b63\u5e38\u8fdb\u7a0b\uff0c\u544a\u8bc9\u8c03\u5ea6\u5668\u4e0d\u8981\u56e0\u8d85\u65f6\u800c\u91cd\u542f\u65b0\u4efb\u52a1"
          },
          "check_point": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Point",
            "description": "(\u672a\u542f\u7528)\u68c0\u67e5\u70b9, \u7528\u4e8e\u8bb0\u5f55\u4efb\u52a1\u7684\u6267\u884c\u8fdb\u5ea6\uff0c\u4ee5\u4fbf\u5728\u4efb\u52a1\u4e2d\u65ad\u540e\u80fd\u591f\u4ece\u68c0\u67e5\u70b9\u6062\u590d, \u5bf9\u4e8e\u67d0\u4e9b\u6ca1\u6709\u5185\u7f6echeckpoint\u529f\u80fd\u7684\u4efb\u52a1\uff0c\u53ef\u4ee5\u4f7f\u7528checkpoint\u7684\u503c\uff0c\u8868\u793a\u4ece\u4ec0\u4e48\u6b65\u9aa4\u7ee7\u7eed\u6267\u884c"
          },
          "runner_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runner Type",
            "description": "\u8fd0\u884c\u5668\u7c7b\u578b"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id",
            "description": "\u7ebf\u7a0bID, \u901a\u5e38\u5bf9\u5e94langgraph \u7684 thread_id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "name",
          "title",
          "description",
          "owner_id"
        ],
        "title": "MtTask"
      },
      "MtTaskStatus": {
        "type": "string",
        "enum": [
          "new",
          "pending",
          "running",
          "success",
          "failed"
        ],
        "title": "MtTaskStatus"
      },
...
      "MttaskUpdateStatusRequest": {
        "properties": {
          "mttask_id": {
            "type": "string",
            "title": "Mttask Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "mttask_id",
          "status"
        ],
        "title": "MttaskUpdateStatusRequest"
      },
...
      "ScheduleCreateRequest": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/MtTaskType",
            "description": "\u4efb\u52a1\u7c7b\u578b"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "\u6807\u9898\u680f"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "\u63cf\u8ff0"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "\u662f\u5426\u6fc0\u6d3b",
            "default": false
          },
          "params": {
            "type": "object",
            "title": "Params",
            "description": "\u4efb\u52a1\u53c2\u6570",
            "default": {}
          },
          "status": {
            "$ref": "#/components/schemas/ScheduleStatus",
            "description": "\u72b6\u6001",
            "default": "inactive"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "task_type"
        ],
        "title": "ScheduleCreateRequest"
      },
      "ScheduleCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ScheduleCreateResponse"
      },
      "ScheduleDetailPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/MtTaskType",
            "description": "\u4efb\u52a1\u7c7b\u578b"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "\u6807\u9898\u680f"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "\u63cf\u8ff0"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "\u662f\u5426\u6fc0\u6d3b",
            "default": false
          },
          "params": {
            "type": "object",
            "title": "Params",
            "description": "\u4efb\u52a1\u53c2\u6570",
            "default": {}
          },
          "status": {
            "$ref": "#/components/schemas/ScheduleStatus",
            "description": "\u72b6\u6001",
            "default": "inactive"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_type"
        ],
        "title": "ScheduleDetailPublic"
      },
      "ScheduleListRequest": {
        "properties": {
          "q": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "active_only": {
            "type": "boolean",
            "title": "Active Only",
            "default": false
          }
        },
        "type": "object",
        "title": "ScheduleListRequest"
      },
...
      "ScheduleUpdateRequest": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/MtTaskType",
            "description": "\u4efb\u52a1\u7c7b\u578b"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "\u6807\u9898\u680f"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "\u63cf\u8ff0"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "\u662f\u5426\u6fc0\u6d3b",
            "default": false
          },
          "params": {
            "type": "object",
            "title": "Params",
            "description": "\u4efb\u52a1\u53c2\u6570",
            "default": {}
          },
          "status": {
            "$ref": "#/components/schemas/ScheduleStatus",
            "description": "\u72b6\u6001",
            "default": "inactive"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_type"
        ],
        "title": "ScheduleUpdateRequest"
      },
      "ScheduleUpdateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ScheduleUpdateResponse"
      },
...
      "SelectItemsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DrowndownSelectItem"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count"
        ],
        "title": "SelectItemsResponse"
      },
...
      "TaskListRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "siteId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siteid"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "title": "TaskListRequest"
      },
      "TaskListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TaskItemPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "TaskListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}